Implement reactive processing of deals file
This is a test approach of processing CSV file in a reactive manner
using Observable collections and Rx library. The goal of this approach
is to get rid of big CSV files caching in memory of the process. With
the first bytes recieved, Observable generates Deal record objects to
subscribers. The Deal objects are pushed to the client immediately using
WebSockets transport implemented by SignalR.
Note: because of significant impact on performance when large
collections are bound, current table output is limited to the first 30
records. In future versions virtual table will be implemented.